Parte 4, ejercicio blogs, eliminacion de blogs solo por el usuario que lo creo

+++ part_4/exercises/blogs/controllers/BlogController.js
@@ -52,9 +52,30 @@ BlogRouter.post('/', async (req, res) => {
 
 // Eliminar un blog
 BlogRouter.delete('/:id', async (req, res) => {
-    await Blog.findByIdAndDelete(req.params.id)
+    // Se decodifica el token mediante el metodo `jwt.token`
+    const decodedToken = jwt.verify(req.token, process.env.SECRET)
 
-    res.status(204).end()
+    if (!decodedToken) return res.status(401).json({ error: 'invalid token' })
+
+    // Se obtiene el blog a eliminar. Mediante `populate` se obtienen los datos del usuario relacionado al blog
+    const blog = await Blog.findById(req.params.id).populate('user')
+
+    if (!blog) return res.status(404).json({ error: 'the provided blog doesn\'t exist' })
+
+    // Se valida que el id del usuario obtenido del token decodificado sea igual al id del usuario asociado al blog
+    if (decodedToken.id === blog.user.id) {
+        // Se elimina el blog de la base de datos
+        const deletedBlog = await Blog.findByIdAndDelete(req.params.id)
+
+        // Se desvincula el blog del usuario que lo ha creado
+        const user = await User.findById(blog.user.id)
+        user.blogs = user.blogs.filter(blog => blog.toString() !== deletedBlog.id)
+        await user.save()
+
+        return res.status(204).end()
+    }
+
+    res.status(401).end()
 })
 
 // Modificar el contenido de un blog existente en la base de datos
